requirements.txt
pyautogui>=0.9.54
Pillow>=10.0.0